requirements.txt
django
djangorestframework
django-allauth
django-rest-auth
django-rest-auth[with_social]
django-redis

boto3
botocore
django-dotenv==1.4.2
django-storages==1.8


django-cors-headers==3.2.1
django-cors-middleware==1.5.0
gunicorn==20.0.4
whitenoise==5.0.1
django-heroku==0.3.1

celery==4.4.6
redis==3.5.3
xlrd
pandas